+++ blueprint/billing-stripe/__test__/test-utils.ts
@@ -12,7 +12,7 @@ import {
   TEST_TOKENS,
   TestSetupResult
 } from '@forklaunch/testing';
-import { EntityManager, MikroORM } from '@mikro-orm/core';
+import { EntityManager } from '@mikro-orm/core';
 import dotenv from 'dotenv';
 import * as path from 'path';
 import { PartyEnum } from '../domain/enum/party.enum';
@@ -90,7 +90,7 @@ export const cleanupTestDatabase = async (): Promise<void> => {
 };
 
 export async function clearDatabase(options?: {
-  orm?: MikroORM;
+  orm?: TestSetupResult['orm'];
   redis?: TestSetupResult['redis'];
 }): Promise<void> {
   await clearTestDatabase(options);