+++ blueprint/messaging-twilio/__test__/test-utils.ts
@@ -7,7 +7,7 @@ import {
   TEST_TOKENS,
   TestSetupResult
 } from '@forklaunch/testing';
-import { EntityManager, MikroORM } from '@mikro-orm/core';
+import { EntityManager } from '@mikro-orm/core';
 import dotenv from 'dotenv';
 import Redis from 'ioredis';
 import * as path from 'path';
@@ -50,7 +50,7 @@ export const cleanupTestDatabase = async (): Promise<void> => {
 };
 
 export const clearDatabase = async (options?: {
-  orm?: MikroORM;
+  orm?: TestSetupResult['orm'];
   redis?: Redis;
 }): Promise<void> => {
   await clearTestDatabase(options);